added invert_block and write_memory functions

+++ ECEN5813_Project3_DK_SE/memtest_functions/memtest.h
@@ -22,7 +22,7 @@ uint8_t *allocate_bytes(uint8_t num_bytes);
 void free_allocated(uint8_t* mem);
 bool verify_memory(uint8_t* mem);
 void display_memory(uint8_t* mem, uint8_t num_bytes);
-void write_memory(uint8_t* mem, uint8_t offset, uint8_t len, uint8_t val);
+void write_memory(uint8_t* mem, uint8_t offset, uint8_t len, uint16_t val);
 void invert_block(uint8_t* mem, uint8_t offset, uint8_t num_bytes);
 void write_pattern(uint8_t* mem, uint8_t num_bytes, uint8_t seed);
 bool compare_pattern(uint8_t* mem, uint8_t num_bytes, uint8_t seed);
